fix: add force flag on compress

+++ xform/src/compress.rs
@@ -15,7 +15,7 @@ use lzw::UnixLZWWriter;
 use plib::PROJECT_NAME;
 use std::fs::{self, File};
 use std::io::{self, Write};
-use std::path::PathBuf;
+use std::path::{Path, PathBuf};
 
 const NAME_MAX: usize = 255;
 
@@ -23,16 +23,16 @@ const NAME_MAX: usize = 255;
 #[derive(Parser, Debug)]
 #[command(author, version, about, long_about)]
 struct Args {
-    /// Write to the standard output; the input file is not changed, and no .Z files are created.
-    #[arg(short = 'c', long)]
-    stdout: bool,
-
     /// Specify the maximum number of bits to use in a code. 9 <= bits <= 14
     #[arg(short = 'b')]
     bits: Option<u32>,
 
+    /// Write to the standard output; the input file is not changed, and no .Z files are created.
+    #[arg(short = 'c', long)]
+    stdout: bool,
+
     /// Do not prompt for overwriting files
-    #[arg(short, long)]
+    #[arg(short = 'f', long)]
     force: bool,
 
     /// Write messages to standard error concerning the expansion of each file.
@@ -43,7 +43,7 @@ struct Args {
     files: Vec<PathBuf>,
 }
 
-fn compress_file(args: &Args, pathname: &PathBuf) -> io::Result<()> {
+fn compress_file(args: &Args, pathname: &PathBuf) -> io::Result<i32> {
     let mut file = plib::io::input_stream(pathname, false)?;
 
     let mut encoder = UnixLZWWriter::new(args.bits);
@@ -56,33 +56,58 @@ fn compress_file(args: &Args, pathname: &PathBuf) -> io::Result<()> {
     out_buf.extend_from_slice(&encoder.close()?);
     let out_buf_size = out_buf.len();
 
-    if args.stdout {
-        io::stdout().write_all(&out_buf)?;
-    } else {
-        let fname = format!("{}.Z", pathname.file_name().unwrap().to_str().unwrap());
-        // If "on adding .Z" the name exceeds the limit of NAME_MAX, then we output on the stdout
-        if fname.len() > NAME_MAX {
+    // check if the output buffer size is greater than input buffer size
+    // this can indicate that the compressed file is greater than the input file
+    //
+    // in such case using '-f' works, i.e "force", which means we don't care if size
+    // isn't reduced, just compress it
+    if (out_buf_size < inp_buf_size) || args.force {
+        if args.stdout {
             io::stdout().write_all(&out_buf)?;
         } else {
-            let mut new_file = pathname.clone();
-            new_file.set_file_name(format!("{fname}"));
-            let mut f = File::create(&new_file)?;
+            let fname = format!("{}.Z", pathname.file_name().unwrap().to_str().unwrap());
+            // If "on adding .Z" the name exceeds the limit of NAME_MAX, then we output on the stdout
+            if fname.len() > NAME_MAX {
+                io::stdout().write_all(&out_buf)?;
+            } else {
+                let path = Path::new(&fname);
+                if path.exists() && !args.force {
+                    print!("Do you want to overwrite {fname} (y)es or (n)o ?");
+                    io::stdout().flush().unwrap();
 
-            fs::remove_file(&pathname)?;
-            f.write_all(&out_buf)?;
+                    let mut inp = String::new();
+                    io::stdin().read_line(&mut inp).unwrap();
 
-            if args.verbose {
-                println!(
-                    "{}: -- replaced with {} Compression: {:.1}%",
-                    pathname.display(),
-                    new_file.display(),
-                    100_f32 - (out_buf_size as f32 / inp_buf_size as f32) * 100_f32
-                );
+                    if !inp.starts_with('y') {
+                        println!("{fname} not overwritten");
+                        return Ok(1);
+                    }
+                }
+
+                let mut new_file = pathname.clone();
+                new_file.set_file_name(format!("{fname}"));
+                let mut f = File::create(&new_file)?;
+
+                fs::remove_file(&pathname)?;
+                f.write_all(&out_buf)?;
+
+                if args.verbose {
+                    println!(
+                        "{}: -- replaced with {} Compression: {:.1}%",
+                        pathname.display(),
+                        new_file.display(),
+                        100_f32 - (out_buf_size as f32 / inp_buf_size as f32) * 100_f32
+                    );
+                }
             }
         }
+    } else {
+        // error status code when file was not compressed as they would have increased in size and (-f
+        // was not provided)
+        return Ok(2);
     }
 
-    Ok(())
+    Ok(0)
 }
 
 fn main() -> Result<(), Box<dyn std::error::Error>> {
@@ -98,9 +123,14 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let mut exit_code = 0;
 
     for filename in &args.files {
-        if let Err(e) = compress_file(&args, filename) {
-            exit_code = 1;
-            eprintln!("{}: {}", filename.display(), e);
+        match compress_file(&args, filename) {
+            Ok(v) => {
+                exit_code = v;
+            }
+            Err(e) => {
+                exit_code = 1;
+                eprintln!("{}: {}", filename.display(), e);
+            }
         }
     }
